修改一些处理逻辑：
* filterStmt 生成时直接进行 Expression bind

* value compare 增加 cast_to 逻辑

*修改 value compare 的返回值，返回枚举类型表示比较结果

*对调用 value compare 的地方进行修改，使结果更加明确

* value 增加了返回 RC 码的 get_value 方法，内部转换，转换失败返回 RC
  错误码

+++ src/observer/sql/expr/tuple.h
@@ -149,7 +149,7 @@ public:
         return rc;
       }
 
-      result = this_value.compare(other_value);
+      result =static_cast<int>(this_value.compare(other_value));
       if (0 != result) {
         return rc;
       }
